+++ doc/conf.py
@@ -81,6 +81,7 @@ linkcheck_ignore = [
     "http://dx.doi.org/10.1088/0965-0393/23/8/083501",  # 403 Client Error: Forbidden for url
     "https://doi.org/10.1111/jmi.12065",  # 403 Client Error: Forbidden for url
     "https://doi.org/10.1111/j.0022-2720.2004.01293.x",  # 403 Client Error: Forbidden for url
+    "https://doi.org/10.1016/j.ultramic.2016.12.021",
 ]
 
 
@@ -88,6 +89,9 @@ linkcheck_ignore = [
 templates_path = [
     "_templates",
 ]
+_version = pyxem.__version__
+version_match = "dev" if "dev" in _version else ".".join(_version.split(".")[:2])
+
 
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files. This image also affects
@@ -103,9 +107,14 @@ html_theme_options = {
     "navigation_with_keys": True,
     "show_toc_level": 2,
     "use_edit_page_button": True,
-    "announcement": "Check out the Free "
-    " <a href='https://www.ntnu.edu/web/temgemini/nordtemhub'>(Big) Data analysis of 4D-STEM</a> "
-    " workshop in Trondheim, Norway, 11-13 June 2024",
+    "announcement": "Check out the "
+    "<a href='https://www.diamond.ac.uk/Home/Events/2024/HyperSpy-Workshop-2024.html'>Hyperspy Workshop</a> "
+    "May 13-17, 2024 Online",
+    "switcher": {
+        "json_url": "https://pyxem.readthedocs.io/en/latest/_static/switcher.json",
+        "version_match": version_match,
+    },
+    "navbar_start": ["navbar-logo", "version-switcher"],
 }
 
 html_context = {