+++ pyxem/tests/utils/test_calibration_utils.py
@@ -90,16 +90,20 @@ class TestCalibrationClass:
             calibration.detector(pixel_size=0.1, detector_distance=1)
         calibration.beam_energy = 200
         calibration.detector(pixel_size=0.1, detector_distance=1)
+        # The center, in pixel coordinates, is (4.5, 4.5)
+        # When using a detector, this gets rounded down to 4
+        assert calibration.center == [4, 4]
         calibration.detector(
             pixel_size=0.1, detector_distance=1, beam_energy=200, units="k_nm^-1"
         )
         assert calibration.flat_ewald is False
+        assert calibration.center == [4, 4]
         with pytest.raises(ValueError):
             calibration(scale=0.01)
         assert calibration.scale is None
         with pytest.raises(ValueError):
             calibration(center=(5, 5))
-        assert calibration.center == [5, 5]
+        assert calibration.center == [4, 4]
 
         with pytest.raises(ValueError):
             calibration.detector(pixel_size=0.1, detector_distance=1, units="nm^-1")
@@ -128,9 +132,9 @@ class TestCalibrationClass:
 
     def test_get_slices_and_factors(self):
         s = Diffraction2D(np.zeros((100, 100)))
-        s.calibrate(scale=0.1, center=None)
-        slices, factors, factor_slices = s.calibrate._get_slices_and_factors(
-            npt=100, npt_azim=360, radial_range=(0, 4)
+        s.calibration(scale=0.1, center=None)
+        slices, factors, factor_slices = s.calibration._get_slices_and_factors(
+            npt=100, npt_azim=360, radial_range=(0, 4), azimuthal_range=(0, 2 * np.pi)
         )
         # check that the number of pixels for each radial slice is the same
         sum_factors = [np.sum(factors[f[0] : f[1]]) for f in factor_slices]
@@ -143,8 +147,8 @@ class TestCalibrationClass:
         # n = npt_azim
         all_sum = np.sum(sum_factors)
         assert np.allclose(all_sum, 3.1415 * 40**2, atol=1)
-        slices, factors, factor_slices = s.calibrate._get_slices_and_factors(
-            npt=100, npt_azim=360, radial_range=(0, 15)
+        slices, factors, factor_slices = s.calibration._get_slices_and_factors(
+            npt=100, npt_azim=360, radial_range=(0, 15), azimuthal_range=(0, 2 * np.pi)
         )
         # check that the number of pixels for each radial slice is the same
         sum_factors = [np.sum(factors[f[0] : f[1]]) for f in factor_slices]
@@ -153,14 +157,12 @@ class TestCalibrationClass:
         # Up to rounding due to the fact that we are actually finding the area of an n-gon where
         # n = npt_azim
         all_sum = np.sum(sum_factors)
-        # For some reason we are missing 1 row/ column of pixels on the edge
-        # of the image so this is 9801 instead of 10000!
-        # assert np.allclose(all_sum, 10000, atol=1)
+        assert np.allclose(all_sum, 10000, atol=1)
 
     def test_get_slices_and_factors1d(self):
         s = Diffraction2D(np.zeros((100, 100)))
-        s.calibrate(scale=0.1, center=None)
-        slices, factors, factor_slices, _ = s.calibrate.get_slices1d(
+        s.calibration(scale=0.1, center=None)
+        slices, factors, factor_slices, _ = s.calibration.get_slices1d(
             100, radial_range=(0, 4)
         )
         # check that the number of pixels for each radial slice is the same